+++ packages/core/src/transforms/__tests__/resize.spec.ts
@@ -1,19 +1,28 @@
-import { resize } from '../resize'
-import { TransformFactoryContext } from '../../types'
-import { applyTransforms } from '../../index'
-import sharp, { Sharp } from 'sharp'
-import { join } from 'path'
 import { toMatchImageSnapshot } from 'jest-image-snapshot'
-import { describe, beforeEach, beforeAll, vi, expect, test } from 'vitest'
+import { join } from 'path'
+import sharp, { Sharp } from 'sharp'
+import { afterAll, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest'
+
+import { applyTransforms } from '../../index'
+import { TransformFactoryContext } from '../../types'
+import { resize } from '../resize'
+import { consoleLogger } from '../../lib/logger'
 
 expect.extend({ toMatchImageSnapshot })
 
-describe('width', () => {
-  let dirCtx: TransformFactoryContext
-  beforeAll(() => {
-    dirCtx = { useParam: vi.fn, warn: vi.fn }
-  })
+let dirCtx: TransformFactoryContext
+beforeAll(() => {
+  dirCtx = { useParam: vi.fn(), logger: consoleLogger }
+  vi.spyOn(dirCtx.logger, 'info')
+})
+beforeEach(() => {
+  vi.resetAllMocks()
+})
+afterAll(() => {
+  vi.restoreAllMocks()
+})
 
+describe('width', () => {
   test('keyword "width"', () => {
     const res = resize({ width: '300' }, dirCtx)
 
@@ -81,11 +90,6 @@ describe('width', () => {
 })
 
 describe('height', () => {
-  let dirCtx: TransformFactoryContext
-  beforeAll(() => {
-    dirCtx = { useParam: vi.fn, warn: vi.fn }
-  })
-
   test('keyword "height"', () => {
     const res = resize({ height: '300' }, dirCtx)
 
@@ -153,11 +157,6 @@ describe('height', () => {
 })
 
 describe('width & height', () => {
-  let dirCtx: TransformFactoryContext
-  beforeAll(() => {
-    dirCtx = { useParam: vi.fn, warn: vi.fn }
-  })
-
   test('keywords "width" & "height"', () => {
     const res = resize({ width: '300', height: '300' }, dirCtx)
 
@@ -232,11 +231,6 @@ describe('width & height', () => {
 })
 
 describe('aspect', () => {
-  let dirCtx: TransformFactoryContext
-  beforeAll(() => {
-    dirCtx = { useParam: vi.fn, warn: vi.fn }
-  })
-
   test('keyword "aspect"', () => {
     const res = resize({ aspect: '16:9' }, dirCtx)
 
@@ -360,7 +354,7 @@ describe('aspect', () => {
     })
 
     test('w/ height', async () => {
-     //@ts-expect-error we know this is safe
+      //@ts-expect-error we know this is safe
       const { image } = await applyTransforms([resize({ aspect: '4:3', height: '75' }, dirCtx)], img)
 
       expect(await image.toBuffer()).toMatchImageSnapshot()
@@ -377,6 +371,206 @@ describe('aspect', () => {
       const { image } = await applyTransforms(
         //@ts-expect-error we know this is safe
         [resize({ aspect: '4:3', height: '300', width: '300' }, dirCtx)],
+        img
+      )
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+  })
+})
+
+describe('withoutEnlargement', () => {
+  test('keyword "withoutEnlargement" w/ dimension', () => {
+    const res = resize({ withoutEnlargement: 'true', width: '300' }, dirCtx)
+
+    expect(res).toBeInstanceOf(Function)
+  })
+
+  test('missing', () => {
+    const res = resize({}, dirCtx)
+
+    expect(res).toBeUndefined()
+  })
+
+  test('true w/ missing dimension', () => {
+    const res = resize({ withoutEnlargement: 'true' }, dirCtx)
+
+    expect(res).toBeUndefined()
+  })
+
+  describe('arguments', () => {
+    test('invalid withoutEnlargement', () => {
+      //@ts-expect-error invalid args
+      const res = resize({ withoutEnlargement: 'invalid', width: '300' }, dirCtx)
+
+      expect(res).toBeUndefined()
+    })
+
+    test('empty', () => {
+      const res = resize({ withoutEnlargement: '', width: '300' }, dirCtx)
+
+      expect(res).toBeInstanceOf(Function)
+    })
+
+    test('true', () => {
+      const res = resize({ withoutEnlargement: 'true', width: '300' }, dirCtx)
+
+      expect(res).toBeInstanceOf(Function)
+    })
+  })
+
+  describe('transform', () => {
+    let img: Sharp
+    beforeEach(() => {
+      img = sharp(join(__dirname, '../../__tests__/__fixtures__/pexels-allec-gomes-5195763.png'))
+    })
+
+    test('w/ multiple dimensions', async () => {
+      //@ts-expect-error we know this is safe
+      const { image } = await applyTransforms([resize({ withoutEnlargement: 'true', width: '300;900' }, dirCtx)], img)
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ width', async () => {
+      //@ts-expect-error we know this is safe
+      const { image } = await applyTransforms([resize({ withoutEnlargement: 'true', width: '300' }, dirCtx)], img)
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ height', async () => {
+      //@ts-expect-error we know this is safe
+      const { image } = await applyTransforms([resize({ withoutEnlargement: 'true', height: '300;900' }, dirCtx)], img)
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ aspect', async () => {
+      const { image } = await applyTransforms(
+        //@ts-expect-error we know this is safe
+        [resize({ withoutEnlargement: 'true', aspect: '4:3' }, dirCtx)],
+        img
+      )
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ width & height', async () => {
+      const { image } = await applyTransforms(
+        //@ts-expect-error we know this is safe
+        [resize({ withoutEnlargement: 'true', height: '300', width: '300' }, dirCtx)],
+        img
+      )
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ width & height & aspect', async () => {
+      const { image } = await applyTransforms(
+        //@ts-expect-error we know this is safe
+        [resize({ withoutEnlargement: 'true', aspect: '4:3', height: '300', width: '300' }, dirCtx)],
+        img
+      )
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+  })
+})
+
+describe('withoutReduction', () => {
+  test('keyword "withoutReduction"', () => {
+    const res = resize({ withoutReduction: 'true', width: '900' }, dirCtx)
+
+    expect(res).toBeInstanceOf(Function)
+  })
+
+  test('missing', () => {
+    const res = resize({}, dirCtx)
+
+    expect(res).toBeUndefined()
+  })
+
+  test('true w/ missing dimension', () => {
+    const res = resize({ withoutReduction: 'true' }, dirCtx)
+
+    expect(res).toBeUndefined()
+  })
+
+  describe('arguments', () => {
+    test('invalid withoutReduction', () => {
+      //@ts-expect-error invalid args
+      const res = resize({ withoutReduction: 'invalid', width: '900' }, dirCtx)
+
+      expect(res).toBeUndefined()
+    })
+
+    test('empty', () => {
+      const res = resize({ withoutReduction: '', width: '900' }, dirCtx)
+
+      expect(res).toBeInstanceOf(Function)
+    })
+
+    test('true', () => {
+      const res = resize({ withoutReduction: 'true', width: '900' }, dirCtx)
+
+      expect(res).toBeInstanceOf(Function)
+    })
+  })
+
+  describe('transform', () => {
+    let img: Sharp
+    beforeEach(() => {
+      img = sharp(join(__dirname, '../../__tests__/__fixtures__/pexels-allec-gomes-5195763.png'))
+    })
+
+    test('w/ multiple dimensions', async () => {
+      //@ts-expect-error we know this is safe
+      const { image } = await applyTransforms([resize({ withoutReduction: 'true', width: '300;900' }, dirCtx)], img)
+
+      expect(dirCtx.logger.info).toHaveBeenCalledOnce()
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ width', async () => {
+      //@ts-expect-error we know this is safe
+      const { image } = await applyTransforms([resize({ withoutReduction: 'true', width: '900' }, dirCtx)], img)
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ height', async () => {
+      //@ts-expect-error we know this is safe
+      const { image } = await applyTransforms([resize({ withoutReduction: 'true', height: '900' }, dirCtx)], img)
+
+      expect(dirCtx.logger.info).toHaveBeenCalledTimes(0)
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ aspect', async () => {
+      const { image } = await applyTransforms(
+        //@ts-expect-error we know this is safe
+        [resize({ withoutReduction: 'true', aspect: '4:3' }, dirCtx)],
+        img
+      )
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ width & height', async () => {
+      const { image } = await applyTransforms(
+        //@ts-expect-error we know this is safe
+        [resize({ withoutReduction: 'true', height: '900', width: '900' }, dirCtx)],
+        img
+      )
+
+      expect(await image.toBuffer()).toMatchImageSnapshot()
+    })
+
+    test('w/ width & height & aspect', async () => {
+      const { image } = await applyTransforms(
+        //@ts-expect-error we know this is safe
+        [resize({ withoutReduction: 'true', aspect: '4:3', height: '900', width: '900' }, dirCtx)],
         img
       )