+++ packages/core/src/transforms/flatten.ts
@@ -6,7 +6,7 @@ export interface FlattenOptions {
   flatten: '' | 'true'
 }
 
-export const flatten: TransformFactory<FlattenOptions> = (config, ctx) => {
+export const flatten: TransformFactory<FlattenOptions> = (config) => {
   if (config.flatten !== '' && config.flatten !== 'true') return
 
   return function flattenTransform(image) {