+++ packages/core/src/types.ts
@@ -2,17 +2,26 @@ import { Sharp } from 'sharp'
 
 export type ImageConfig = Record<string, unknown>
 
-export interface TransformFactoryContext {
-  useParam: (parameter: string) => void
-  warn: (message: string) => void
+export interface Logger {
+  info: (msg: string) => void
+  warn: (msg: string) => void
+  error: (msg: string) => void
 }
 
-export type TransformFactory<A = Record<string,unknown>> = (
+export interface TransformFactoryContext {
+  useParam: (parameter: string) => void
+  logger: Logger
+}
+
+export type TransformFactory<A = Record<string, unknown>> = (
   metadata: Partial<ImageConfig & A>,
   ctx: TransformFactoryContext
 ) => ImageTransformation | undefined
 
-export type TransformOption<A = Record<string,unknown>, T = unknown> = (metadata: Partial<ImageConfig & A>, image: Sharp) => T | undefined
+export type TransformOption<A = Record<string, unknown>, T = unknown> = (
+  metadata: Partial<ImageConfig & A>,
+  image: Sharp
+) => T | undefined
 
 export type ImageTransformation = (image: Sharp) => Sharp | Promise<Sharp>
 
@@ -32,18 +41,18 @@ export type OutputFormat = (args?: string[]) => (metadata: ImageConfig[]) => unk
  * choose between width descriptor, pixel density descriptor, or no descriptor.
  */
 export interface Source {
-  src: string;
-  w: number;
+  src: string
+  w: number
 }
 
 /**
  * The picture output format.
  */
 export interface Picture {
-  sources: Record<string, Source[]>;
+  sources: Record<string, Source[]>
   fallback: {
-    src: string;
-    w: number;
-    h: number;
-  };
+    src: string
+    w: number
+    h: number
+  }
 }