+++ packages/core/src/vitest.config.ts
@@ -3,5 +3,5 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     setupFiles: ['../../test-setup.js']
-  },
+  }
 })