+++ packages/vite/src/__tests__/main.test.ts
@@ -1,4 +1,4 @@
-import { build } from 'vite'
+import { build, createLogger } from 'vite'
 import { imagetools } from '../index'
 import { join } from 'path'
 import { getFiles, testEntry } from './util'
@@ -6,9 +6,14 @@ import { toMatchImageSnapshot } from 'jest-image-snapshot'
 import { OutputAsset, OutputChunk, RollupOutput } from 'rollup'
 import { JSDOM } from 'jsdom'
 import sharp from 'sharp'
-import { describe, test, expect, it } from 'vitest'
+import { afterEach, describe, test, expect, it, vi } from 'vitest'
+import { createBasePath } from '../utils'
 
 expect.extend({ toMatchImageSnapshot })
+
+afterEach(() => {
+  vi.restoreAllMocks()
+})
 
 describe('vite-imagetools', () => {
   describe('options', () => {
@@ -208,10 +213,92 @@ describe('vite-imagetools', () => {
       })
     })
 
-    describe('silent', () => {
-      test('false by default', () => {})
-      test('true disables all warnings', () => {})
-      test('false enables warnings', () => {})
+    describe('logging', () => {
+      test('logs info messages to console', async () => {
+        const logger = createLogger('info')
+        const spy = vi.spyOn(logger, 'info')
+        await build({
+          root: join(__dirname, '__fixtures__'),
+          logLevel: 'info',
+          customLogger: logger,
+          build: { write: false },
+          plugins: [
+            testEntry(`
+                          import Image from "./with-metadata.png?warn"
+                          window.__IMAGE__ = Image
+                      `),
+            imagetools({
+              extendTransforms() {
+                return [
+                  (config, context) => {
+                    context.logger.info('An info message')
+                    return (image) => image
+                  }
+                ]
+              }
+            })
+          ]
+        })
+
+        expect(spy).toHaveBeenCalledWith('An info message')
+      })
+      test('logs warn messages through rollup', async () => {
+        const logger = createLogger('info')
+        const spy = vi.spyOn(logger, 'warn')
+        await build({
+          root: join(__dirname, '__fixtures__'),
+          logLevel: 'warn',
+          customLogger: logger,
+          build: { write: false },
+          plugins: [
+            testEntry(`
+                          import Image from "./with-metadata.png?warn"
+                          window.__IMAGE__ = Image
+                      `),
+            imagetools({
+              extendTransforms() {
+                return [
+                  (config, context) => {
+                    context.logger.warn('A warning')
+                    return (image) => image
+                  }
+                ]
+              }
+            })
+          ]
+        })
+
+        expect(spy.mock.lastCall?.[0]).toContain('A warning')
+      })
+      test('logs error messages through rollup', async () => {
+        try {
+          await build({
+            root: join(__dirname, '__fixtures__'),
+            logLevel: 'warn',
+            build: { write: false },
+            plugins: [
+              testEntry(`
+                            import Image from "./with-metadata.png?warn"
+                            window.__IMAGE__ = Image
+                        `),
+              imagetools({
+                extendTransforms() {
+                  return [
+                    (config, context) => {
+                      context.logger.error('An error')
+                      return (image) => image
+                    }
+                  ]
+                }
+              })
+            ]
+          })
+          fail()
+        } catch (err: any) {
+          expect(err.plugin).toEqual('imagetools')
+          expect(err.message).toMatch(/An error$/)
+        }
+      })
     })
 
     describe('removeMetadata', () => {
@@ -557,7 +644,17 @@ describe('vite-imagetools', () => {
     const { window } = new JSDOM(``, { runScripts: 'outside-only' })
     window.eval(files[0].code)
 
-     expect(window.__IMAGE__).toBe('/assets/with-metadata-404f605d.png 600w')
+    expect(window.__IMAGE__).toBe('/assets/with-metadata-404f605d.png 600w')
   })
 
+  describe('utils', () => {
+    test('createBasePath', () => {
+      expect(createBasePath('')).toBe('/@imagetools/')
+      expect(createBasePath('/')).toBe('/@imagetools/')
+      expect(createBasePath('/base')).toBe('/base/@imagetools/')
+      expect(createBasePath('/base/')).toBe('/base/@imagetools/')
+      expect(createBasePath('http://localhost:9000/frontend')).toBe('http://localhost:9000/frontend/@imagetools/')
+      expect(createBasePath('http://localhost:9000/frontend/')).toBe('http://localhost:9000/frontend/@imagetools/')
+    })
+  })
 })